requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
reportlab==4.0.8
python-multipart==0.0.6 